+++ crates/gitbutler-branch-actions/Cargo.toml
@@ -29,7 +29,6 @@ gitbutler-operating-modes.workspace = true
 gitbutler-cherry-pick.workspace = true
 gitbutler-oxidize.workspace = true
 gitbutler-stack.workspace = true
-gitbutler-stack-api.workspace = true
 gitbutler-patch-reference.workspace = true
 serde = { workspace = true, features = ["std"] }
 bstr.workspace = true
