+++ crates/gitbutler-branch-actions/src/branch_manager/branch_creation.rs
@@ -13,8 +13,8 @@ use gitbutler_repo::{
     LogUntil, RepositoryExt,
 };
 use gitbutler_repo_actions::RepoActionsExt;
+use gitbutler_stack::StackExt;
 use gitbutler_stack::{BranchOwnershipClaims, Stack, StackId};
-use gitbutler_stack_api::StackExt;
 use gitbutler_time::time::now_since_unix_epoch_ms;
 use tracing::instrument;
