+++ crates/gitbutler-branch-actions/src/integration.rs
@@ -12,8 +12,8 @@ use gitbutler_operating_modes::OPEN_WORKSPACE_REFS;
 use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_repo::SignaturePurpose;
 use gitbutler_repo::{LogUntil, RepositoryExt};
+use gitbutler_stack::StackExt;
 use gitbutler_stack::{Stack, VirtualBranchesHandle};
-use gitbutler_stack_api::StackExt;
 use tracing::instrument;
 
 use crate::{branch_manager::BranchManagerExt, conflicts, VirtualBranchesExt};
