+++ crates/gitbutler-branch-actions/src/move_commits.rs
@@ -7,8 +7,8 @@ use gitbutler_command_context::CommandContext;
 use gitbutler_commit::commit_ext::CommitExt;
 use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_repo::{rebase::cherry_rebase_group, LogUntil, RepositoryExt};
+use gitbutler_stack::StackExt;
 use gitbutler_stack::{OwnershipClaim, StackId};
-use gitbutler_stack_api::StackExt;
 use std::collections::HashMap;
 
 /// moves commit from the branch it's in to the top of the target branch
