+++ crates/gitbutler-branch-actions/src/reorder_commits.rs
@@ -2,8 +2,8 @@ use anyhow::{bail, Context as _, Result};
 use gitbutler_command_context::CommandContext;
 use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_repo::{rebase::cherry_rebase_group, LogUntil, RepositoryExt as _};
+use gitbutler_stack::StackExt;
 use gitbutler_stack::StackId;
-use gitbutler_stack_api::StackExt;
 
 use crate::{
     branch_trees::{
