+++ crates/gitbutler-branch-actions/src/undo_commit.rs
@@ -2,8 +2,8 @@ use anyhow::{bail, Context as _, Result};
 use gitbutler_command_context::CommandContext;
 use gitbutler_commit::commit_ext::CommitExt as _;
 use gitbutler_repo::{rebase::cherry_rebase_group, LogUntil, RepositoryExt as _};
+use gitbutler_stack::StackExt;
 use gitbutler_stack::{Stack, StackId};
-use gitbutler_stack_api::StackExt;
 
 use crate::VirtualBranchesExt as _;
