--- crates/gitbutler-stack-api/src/lib.rs
@@ -1,7 +1,0 @@
-mod heads;
-mod series;
-mod stack_ext;
-pub use series::Series;
-pub use stack_ext::{
-    commit_by_oid_or_change_id, CommitsForId, PatchReferenceUpdate, StackExt, TargetUpdate,
-};
