+++ crates/gitbutler-stack/src/heads.rs
