+++ crates/gitbutler-stack/src/lib.rs
@@ -9,3 +9,11 @@ pub use ownership::{reconcile_claims, BranchOwnershipClaims, ClaimOutcome};
 pub use stack::{Stack, StackId};
 pub use state::{VirtualBranches as VirtualBranchesState, VirtualBranchesHandle};
 pub use target::Target;
+
+mod heads;
+mod series;
+mod stack_ext;
+pub use series::Series;
+pub use stack_ext::{
+    commit_by_oid_or_change_id, CommitsForId, PatchReferenceUpdate, StackExt, TargetUpdate,
+};
