+++ crates/gitbutler-stack/tests/fixtures/stacking.sh
